backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pydantic==2.8.2
pyserial==3.5
pymongo==4.8.0
langchain==0.2.14
langchain-ollama==0.1.0
python-dotenv==1.0.1
